+++ haze/src/commonMain/kotlin/dev/chrisbanes/haze/TrimMemoryCallback.kt
@@ -3,9 +3,27 @@
 
 package dev.chrisbanes.haze
 
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
 import kotlinx.coroutines.DisposableHandle
 
 internal expect fun registerTrimMemoryCallback(
   context: PlatformContext,
+  lifecycle: Lifecycle?,
   callback: (TrimMemoryLevel) -> Unit,
 ): DisposableHandle
+
+internal fun registerLifecycleTrimMemoryCallback(
+  lifecycle: Lifecycle?,
+  callback: (TrimMemoryLevel) -> Unit,
+): DisposableHandle {
+  if (lifecycle == null) return DisposableHandle {}
+
+  val observer = LifecycleEventObserver { _, event ->
+    if (event == Lifecycle.Event.ON_STOP) {
+      callback(TrimMemoryLevel.MODERATE)
+    }
+  }
+  lifecycle.addObserver(observer)
+  return DisposableHandle { lifecycle.removeObserver(observer) }
+}